Moved realtime and time to combined package
