--- src/main/java/org/betonquest/betonquest/quest/condition/realtime/package-info.java
@@ -1,5 +1,0 @@
-/**
- * {@link org.betonquest.betonquest.api.quest.condition.PlayerlessCondition PlayerlessCondition}
- * implementations of real time conditions.
- */
-package org.betonquest.betonquest.quest.condition.realtime;
