+++ src/main/java/org/betonquest/betonquest/quest/registry/CoreQuestTypes.java
@@ -74,15 +74,15 @@ import org.betonquest.betonquest.quest.condition.objective.ObjectiveConditionFac
 import org.betonquest.betonquest.quest.condition.point.GlobalPointConditionFactory;
 import org.betonquest.betonquest.quest.condition.point.PointConditionFactory;
 import org.betonquest.betonquest.quest.condition.random.RandomConditionFactory;
-import org.betonquest.betonquest.quest.condition.realtime.DayOfWeekConditionFactory;
-import org.betonquest.betonquest.quest.condition.realtime.PartialDateConditionFactory;
-import org.betonquest.betonquest.quest.condition.realtime.RealTimeConditionFactory;
 import org.betonquest.betonquest.quest.condition.ride.RideConditionFactory;
 import org.betonquest.betonquest.quest.condition.scoreboard.ScoreboardConditionFactory;
 import org.betonquest.betonquest.quest.condition.sneak.SneakConditionFactory;
 import org.betonquest.betonquest.quest.condition.tag.GlobalTagConditionFactory;
 import org.betonquest.betonquest.quest.condition.tag.TagConditionFactory;
-import org.betonquest.betonquest.quest.condition.time.TimeConditionFactory;
+import org.betonquest.betonquest.quest.condition.time.ingame.TimeConditionFactory;
+import org.betonquest.betonquest.quest.condition.time.real.DayOfWeekConditionFactory;
+import org.betonquest.betonquest.quest.condition.time.real.PartialDateConditionFactory;
+import org.betonquest.betonquest.quest.condition.time.real.RealTimeConditionFactory;
 import org.betonquest.betonquest.quest.condition.variable.VariableConditionFactory;
 import org.betonquest.betonquest.quest.condition.weather.WeatherConditionFactory;
 import org.betonquest.betonquest.quest.condition.world.WorldConditionFactory;
